Split SlimLinkedOCEL binding functions, fix SQL
