+++ process_mining/src/analysis/object_centric/mod.rs
@@ -1,3 +1,4 @@
 //! Object-centric Process Analysis
 
 pub mod object_attribute_changes;
+pub mod oc_performance;
